feat: Complete trading system with position sizing fixes, AI integration, and comprehensive documentation
Major Features:
- Fixed position sizing with 10% max per position cap
- Dynamic stop loss and take profit with bracket orders
- Multi-indicator strategy (EMA crossover with volume confirmation)
- AI-powered copilot with Perplexity and OpenRouter integration
- Command system for portfolio management
- Real-time streaming and WebSocket support
- Comprehensive risk management and circuit breakers

Bug Fixes:
- Fixed database schema (added prev_ema columns)
- Fixed volume data type (int instead of float)
- Fixed position sizing calculation with min stop distance
- Fixed order rejection by capping position size instead of rejecting

Documentation:
- GuideForDummies.md - Complete beginner's guide
- ENHANCEMENT_PROPOSAL_MAXIMIZE_PROFITS.md - Future roadmap
- Multiple troubleshooting and setup guides
- Architecture and security documentation

Testing:
- Added diagnostic tools (diagnose_trading.py, test_signal_detection.py)
- Multiple test modules for all major components
- Paper trading verified and working

Current Status:
- System is live and taking positions
- MSFT, META, NVDA positions active with stops/targets
- All safety features operational
- Ready for Phase 1 enhancements

+++ backend/data/features.py
@@ -70,7 +70,7 @@ class FeatureEngine:
                 'prev_ema_short': float(prev['ema_short']),
                 'prev_ema_long': float(prev['ema_long']),
                 'atr': float(latest['atr']),
-                'volume': float(latest['volume']),
+                'volume': int(latest['volume']),  # Convert to int for BIGINT column
                 'volume_zscore': FeatureEngine.calculate_volume_zscore(df['volume']),
                 'ema_diff': float(latest['ema_short'] - latest['ema_long']),
                 'ema_diff_pct': float((latest['ema_short'] / latest['ema_long'] - 1) * 100),
